Refactor: Implement collectLatest and flatMapConcat operators on Flow
The `collectFlow` function has been refactored to implement the `collectLatest` and `flatMapConcat` operators on different flows. A new flow has been added that emits food items with delays. The current implementation is to consume and print each food item and the state. Previous implementations using fold, filter, map, onEach, count operators were commented out. The old flow `flow1` that used to output 1 and 2 is commented out. `flatMapConcat` operation is commented out but shows how to use it, it will receive a value and call a function that will return a flow. The old implementation that collects the flow is also commented out.