Refactor: Implement buffer operator on Flow
The `collectFlow` function now utilizes the `buffer` operator to improve the efficiency of flow emissions. `buffer` operator is added to the flow after the `onEach` operation to allow the flow to emit values into a buffer without waiting for the `collect` function to process each item sequentially, this make the process faster.